refactor(cli): split command families
